move pychiquito.rs content to python/mod.rs
