+++ src/frontend/mod.rs
@@ -1,3 +1,2 @@
 pub mod dsl;
-pub mod pychiquito;
 pub mod python;
